renamed cmd/ drivers as prefixcmd
