+++ ironway.go
@@ -3,14 +3,14 @@ package main
 import (
 	"os"
 
-	"github.com/UnnecessaryRain/ironway-core/cmd/gamecommand"
-	"github.com/UnnecessaryRain/ironway-core/cmd/servecommand"
+	"github.com/UnnecessaryRain/ironway-core/cmd/gamecmd"
+	"github.com/UnnecessaryRain/ironway-core/cmd/servecmd"
 	kingpin "gopkg.in/alecthomas/kingpin.v2"
 )
 
 func main() {
 	app := kingpin.New("ironway", "ironway core server")
-	servecommand.Configure(app)
-	gamecommand.Configure(app)
+	servecmd.Configure(app)
+	gamecmd.Configure(app)
 	kingpin.MustParse(app.Parse(os.Args[1:]))
 }
